backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
python-multipart>=0.0.6
google-adk>=0.1.0
python-dotenv>=1.0.0
deprecated>=1.3.0
